Break code into modules, add QoL features
The way we process the list of files to download changed.
Previously, we downloaded files post by post, which was
inefficient due to the inconsistent number of files per post.
Now all file URLs are collected and then processed by
a limited number of concurrent tasks (currently 32).

other new features:
- improved usage hint
- improved message formatting and priority
- use of temporary files to avoid incomplete downloads on error
- nice formatting for number of posts/files, file sizes and delays
- download statistics
- file download message includes file size
- socks5 proxy support
- delay between API requests to avoid rate-limit
- API rate-limit detection, wait on hit